+++ src/collider_component.h
@@ -26,7 +26,7 @@ namespace Radix
 
 			void initialize() override;
 
-			void update(float) override;
+			void update(double) override;
 	};
 }
